refactor(gui-agent): split BrowserControlRenderer into focused components
- Extract StrategySwitch for screenshot display mode controls
- Create MouseCursor component for cursor visualization
- Add ScreenshotDisplay for image rendering logic
- Implement OperationDetailsCard for action details
- Add useMousePosition hook for coordinate handling
- Add useScreenshots hook for image discovery
- Extract data parsing to utility function
- Maintain all existing functionality and behavior
